requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.16.0,<3.17.0
psycopg[c]==3.2.3
django-cors-headers==4.6.0
djangorestframework-simplejwt
django-filter>=24.1,<25.0
dj-rest-auth[with_social]>=2.2.5,<3.0
django-allauth>=0.50.0,<0.53.0
google-auth>=2.29.0,<3.0.0
tensorflow-cpu>=2.12,<3.0
scikit-learn>=1.3,<2.0
pandas>=2.2,<3.0
numpy>=1.26,<2.0
nltk>=3.8,<4.0
joblib>=1.3,<2.0